Clean SQL DDL classes
