+++ core/src/main/java/ch/unibas/dmi/dbis/polyphenydb/sql/ddl/SqlAlterTable.java
@@ -26,26 +26,10 @@
 package ch.unibas.dmi.dbis.polyphenydb.sql.ddl;
 
 
-import static ch.unibas.dmi.dbis.polyphenydb.util.Static.RESOURCE;
-
-import ch.unibas.dmi.dbis.polyphenydb.Transaction;
-import ch.unibas.dmi.dbis.polyphenydb.catalog.entity.CatalogColumn;
-import ch.unibas.dmi.dbis.polyphenydb.catalog.entity.CatalogTable;
-import ch.unibas.dmi.dbis.polyphenydb.catalog.entity.combined.CatalogCombinedTable;
-import ch.unibas.dmi.dbis.polyphenydb.catalog.exceptions.GenericCatalogException;
-import ch.unibas.dmi.dbis.polyphenydb.catalog.exceptions.UnknownCollationException;
-import ch.unibas.dmi.dbis.polyphenydb.catalog.exceptions.UnknownColumnException;
-import ch.unibas.dmi.dbis.polyphenydb.catalog.exceptions.UnknownDatabaseException;
-import ch.unibas.dmi.dbis.polyphenydb.catalog.exceptions.UnknownSchemaException;
-import ch.unibas.dmi.dbis.polyphenydb.catalog.exceptions.UnknownSchemaTypeException;
-import ch.unibas.dmi.dbis.polyphenydb.catalog.exceptions.UnknownTableException;
-import ch.unibas.dmi.dbis.polyphenydb.jdbc.Context;
 import ch.unibas.dmi.dbis.polyphenydb.sql.SqlAlter;
-import ch.unibas.dmi.dbis.polyphenydb.sql.SqlIdentifier;
 import ch.unibas.dmi.dbis.polyphenydb.sql.SqlKind;
 import ch.unibas.dmi.dbis.polyphenydb.sql.SqlOperator;
 import ch.unibas.dmi.dbis.polyphenydb.sql.SqlSpecialOperator;
-import ch.unibas.dmi.dbis.polyphenydb.sql.SqlUtil;
 import ch.unibas.dmi.dbis.polyphenydb.sql.parser.SqlParserPos;
 
 
@@ -66,55 +50,7 @@ public abstract class SqlAlterTable extends SqlAlter {
     }
 
 
-    protected CatalogTable getCatalogTable( Context context, Transaction transaction, SqlIdentifier tableName ) {
-        CatalogTable catalogTable;
-        try {
-            long schemaId;
-            String tableOldName;
-            if ( tableName.names.size() == 3 ) { // DatabaseName.SchemaName.TableName
-                schemaId = transaction.getCatalog().getSchema( tableName.names.get( 0 ), tableName.names.get( 1 ) ).id;
-                tableOldName = tableName.names.get( 2 );
-            } else if ( tableName.names.size() == 2 ) { // SchemaName.TableName
-                schemaId = transaction.getCatalog().getSchema( context.getDatabaseId(), tableName.names.get( 0 ) ).id;
-                tableOldName = tableName.names.get( 1 );
-            } else { // TableName
-                schemaId = transaction.getCatalog().getSchema( context.getDatabaseId(), context.getDefaultSchemaName() ).id;
-                tableOldName = tableName.names.get( 0 );
-            }
-            catalogTable = transaction.getCatalog().getTable( schemaId, tableOldName );
-        } catch ( UnknownDatabaseException e ) {
-            throw SqlUtil.newContextException( tableName.getParserPosition(), RESOURCE.databaseNotFound( tableName.toString() ) );
-        } catch ( UnknownSchemaException e ) {
-            throw SqlUtil.newContextException( tableName.getParserPosition(), RESOURCE.schemaNotFound( tableName.toString() ) );
-        } catch ( UnknownTableException e ) {
-            throw SqlUtil.newContextException( tableName.getParserPosition(), RESOURCE.tableNotFound( tableName.toString() ) );
-        } catch ( UnknownCollationException | UnknownSchemaTypeException | GenericCatalogException e ) {
-            throw new RuntimeException( e );
-        }
-        return catalogTable;
-    }
 
-
-    protected CatalogCombinedTable getCatalogCombinedTable( Context context, Transaction transaction, SqlIdentifier tableName ) {
-        try {
-            return transaction.getCatalog().getCombinedTable( getCatalogTable( context, transaction, tableName ).id );
-        } catch ( GenericCatalogException | UnknownTableException e ) {
-            throw new RuntimeException( e );
-        }
-    }
-
-
-    protected CatalogColumn getCatalogColumn( Context context, Transaction transaction, long tableId, SqlIdentifier columnName ) {
-        CatalogColumn catalogColumn;
-        try {
-            catalogColumn = transaction.getCatalog().getColumn( tableId, columnName.getSimple() );
-        } catch ( GenericCatalogException e ) {
-            throw new RuntimeException( e );
-        } catch ( UnknownColumnException e ) {
-            throw SqlUtil.newContextException( columnName.getParserPosition(), RESOURCE.columnNotFound( columnName.getSimple() ) );
-        }
-        return catalogColumn;
-    }
 
 }
